# Replace synchronized block with ReentrantLock to avoid possible thread pinning with virtual threads (java 21)