It's recommended to avoid using synchronized block with VirtualThreads. Spring Boot starting from 3.2 and other frameworks now allow to run using virtual threads, so it's small, but valuable change for such scenarios.