@mweirauch 2024-09-08
Just for the record:

i7-6820HQ, Fedora 40, 8.0.422-tem; JMX 1.33:

(11.0.24-tem, 17.0.12-tem and 21.0.4-tem with pretty much identical results)

```text
# before

Benchmark                                                          Mode  Cnt        Score         Error   Units
ProcfsStatusBenchmark.collectAverage                               avgt    5       21,530 ±       0,330   ns/op
ProcfsStatusBenchmark.collectAverageContended                      avgt    5       22,660 ±       0,151   ns/op
ProcfsStatusBenchmark.collectSingle                                  ss    5  6020898,400 ± 4289315,078   ns/op

# after

Benchmark                                                          Mode  Cnt        Score         Error   Units
ProcfsStatusBenchmark.collectAverage                               avgt    5       21,517 ±       0,192   ns/op
ProcfsStatusBenchmark.collectAverageContended                      avgt    5       22,576 ±       0,224   ns/op
ProcfsStatusBenchmark.collectSingle                                  ss    5  5290317,400 ± 2752044,273   ns/op
```